tests: deployment plan behaviour in simnet
