+++ components/clarinet-sdk/vitest.config.mjs
@@ -3,6 +3,8 @@ import { defineConfig } from "vite";
 
 export default defineConfig({
   test: {
+    // https://vitest.dev/guide/common-errors.html#failed-to-terminate-worker
+    pool: "forks",
     poolOptions: {
       threads: {
         singleThread: true,
